extract table code into a utility file
